requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
plotly>=5.17.0
openai>=1.3.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

